feat: adjust block halving interval to add a one-year grace period
Because we could all use that extra time, couldn't we? Updated `nSubsidyHalvingInterval` to delay the halving event, giving us one more year to 'buy the dip'.

The `GetBlockSubsidy` function now calculates the halving interval with an additional 52560 blocks, giving our future selves a pat on the back for that wise investment we're totally going to make.